relocated components into separate folders
