+++ components/buttons/EditImageIcon.js
@@ -1,7 +1,7 @@
 import { StyleSheet, TouchableOpacity, Image, View } from 'react-native';
 import { MaterialIcons } from '@expo/vector-icons';
 
-import colors from '../styles/colors';
+import colors from '../../styles/colors';
 
 export default function EditImageIcon({ image, action }) {
   return (
