+++ components/buttons/ProfileButton.js
@@ -1,6 +1,6 @@
 import { View, StyleSheet, Text, TouchableOpacity } from 'react-native';
-import colors from '../styles/colors';
-import textStyle from '../styles/text';
+import colors from '../../styles/colors';
+import textStyle from '../../styles/text';
 import { Entypo } from '@expo/vector-icons'; 
 
 export default function ProfileButton({ title, color, action, style }) {
