+++ screens/SearchResults.js
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from 'react'
 import { StyleSheet, Text, SafeAreaView, View, FlatList, ActivityIndicator, Image } from 'react-native';
 import { getRequestHeaders } from '../utils/api';
 
-import BackButton from '../components/BackButton';
-import Listing from '../components/Listing';
-import EmptyList from '../components/EmptyList';
+import BackButton from '../components/buttons/BackButton';
+import Listing from '../components/cards/Listing';
+import EmptyList from '../components/cards/EmptyList';
 
 import textStyle from '../styles/text';
 import colors from '../styles/colors';
